Fixing BackgroundLayout writing for RSDKv3 and RSDKv4.

+++ RSDKv3/BackgroundLayout.cs
@@ -15,7 +15,7 @@ namespace RSDKv3
             /// <summary>
             /// How fast the line moves without the player moving
             /// </summary>
-            public short ConstantSpeed;
+            public byte ConstantSpeed;
             /// <summary>
             /// a special byte that tells the game what "behaviour" property the layer has
             /// </summary>
@@ -28,7 +28,7 @@ namespace RSDKv3
                 Behaviour = 0;
             }
 
-            public ScrollInfo(byte r, byte c, byte b)
+            public ScrollInfo(short r, byte c, byte b)
             {
                 RelativeSpeed = r;
                 ConstantSpeed = c;
@@ -80,7 +80,7 @@ namespace RSDKv3
             /// <summary>
             /// how fast the layer moves while the player isn't moving
             /// </summary>
-            public short ConstantSpeed;
+            public byte ConstantSpeed;
             /// <summary>
             /// a list of Line positions
             /// </summary>
